% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_migration_risk.R
\name{classify_migration_risk}
\alias{classify_migration_risk}
\title{Classify migration risk based on check and test results}
\usage{
classify_migration_risk(baseline_results, post_results)
}
\arguments{
\item{baseline_results}{data.table from baseline check}

\item{post_results}{data.table from post-migration check}
}
\value{
A list
}
\description{
Classify migration risk based on check and test results
}
