% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_libraries.R
\name{compare_libraries}
\alias{compare_libraries}
\title{Compare two package libraries}
\usage{
compare_libraries(source_pkgs, target_pkgs)
}
\arguments{
\item{source_pkgs}{data.table from \code{list_packages}}

\item{target_pkgs}{data.table from \code{list_packages}}
}
\value{
A list of data.tables and a summary data.frame
}
\description{
Compare two package libraries
}
