% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate_simulation.R
\name{decorate_simulation}
\alias{decorate_simulation}
\title{add send_message function to a user-defined function}
\usage{
decorate_simulation(.f, .port)
}
\arguments{
\item{.f}{function}

\item{.port}{port send_message will communication on}
}
\description{
add send_message function to a user-defined function
}
\details{
decorate_simulation will give the ability of the decorated function
to use a function with a signature courier_msg(x) where
x is a string message. This message will be send via zmq
to the receiving server. Client connection and teardown
happen transparently
}
\examples{
\dontrun{
 PORT <- 5678 # number would be taken from run_receiving_server()
 my_sim <- decorate_simulation(function(rep) {
 courier_msg(paste("on rep", rep))
 # emulate 1 second of simulation 'work'
 Sys.sleep(1)
 return(runif(1, 0, 100))
 }, PORT)

 lapply(1:10, my_sim)
}
}

