% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_install_source.R
\name{detect_install_source}
\alias{detect_install_source}
\title{Generate a pak specification for a package}
\usage{
detect_install_source(
  package,
  version = NULL,
  source_hint = NULL,
  github_ref = NULL
)
}
\arguments{
\item{package}{Package name}

\item{version}{Optional version constraint or exact version}

\item{source_hint}{Optional hint: "CRAN", "Bioconductor", "GitHub", "local"}

\item{github_ref}{Optional GitHub ref like "owner/repo@ref"}
}
\value{
A character vector of pak specs
}
\description{
Generate a pak specification for a package
}
