% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_project_type.R
\name{detect_project_type}
\alias{detect_project_type}
\title{Detect project characteristics}
\usage{
detect_project_type(project_path)
}
\arguments{
\item{project_path}{Path to the project}
}
\value{
A named list
}
\description{
Detect project characteristics
}
