% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_r_installations.R
\name{detect_r_installations}
\alias{detect_r_installations}
\title{Detect R installations on the system}
\usage{
detect_r_installations(search_paths = NULL)
}
\arguments{
\item{search_paths}{Optional extra paths to search}
}
\value{
A data.frame of R installations
}
\description{
Detect R installations on the system
}
