% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_addresses.R
\docType{data}
\name{.pkgglobalenv}
\alias{.pkgglobalenv}
\title{Process address in a database}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.pkgglobalenv
}
\arguments{
\item{dbname}{Name of the database. Default is \code{NULL}.}

\item{host}{Name of the host IP address. Default is \code{NULL}.}

\item{user}{Username. Default is \code{NULL}.}

\item{port}{Port number, typically 5432. Default is \code{NULL}.}

\item{table_name}{Name of the table you want to query data}

\item{slackr_channel}{By default, loaded via environmental variable \code{SLACKR_CHANNEL}.}

\item{slackr_bot}{By default, loaded via environmental variable \code{SLACKR_BOT}.}

\item{slackr_token}{By default, loaded via environmental variable \code{SLACKR_TOKEN}.}

\item{slackr_webhook}{By default, loaded via environmental variable \code{SLACKR_WEBHOOK}.}

\item{verbose}{Include logging to the console. Default \code{FALSE}.}
}
\description{
A main worker function for fetching addresses from a database in chunks,
processing them with Placekey's bulk API and reporting progress via Slack.
}
\keyword{datasets}
