% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_db_data.R
\name{fetch_db_data}
\alias{fetch_db_data}
\title{Retrieve a chunk of data}
\usage{
fetch_db_data(
  dbname = NULL,
  host = NULL,
  user = NULL,
  port = NULL,
  table_name,
  order = 1,
  limit = 10000,
  offset = 0
)
}
\arguments{
\item{dbname}{Name of the database. Default is \code{NULL}.}

\item{host}{Name of the host IP address. Default is \code{NULL}.}

\item{user}{Username. Default is \code{NULL}.}

\item{port}{Port number, typically 5432. Default is \code{NULL}.}

\item{table_name}{Name of the table you want to query data}

\item{order}{Default is \code{1}, or the data's first column.}

\item{limit}{Default is \code{10000}, consistent with Placekey bulk API rate limiting.}

\item{offset}{Number of rows to skip in the data. Default is \code{0}.}
}
\value{
A tibble chunked from the specified database table.
}
\description{
Use Postgres to access a specified database and retrieve a chunk of data
in a specified size as a \code{dplyr} tibble.
}
\examples{
\dontrun{
fetch_db_data(table_name = 'mtcars', offset = 5)
fetch_db_data('exampledb', '123.456.7.89', 'user', 'mtcars')
}
}
