% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ship.R
\name{find_target_lib}
\alias{find_target_lib}
\title{Resolve the primary target library for an R installation}
\usage{
find_target_lib(target_path)
}
\arguments{
\item{target_path}{Full path to target \code{Rscript}.}
}
\value{
Character scalar path to \code{.libPaths()[1L]} in the target R.
}
\description{
Resolve the primary target library for an R installation
}
\keyword{internal}
