% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_placekeys.R
\name{get_placekeys}
\alias{get_placekeys}
\title{Get placekeys using the bulk API}
\usage{
get_placekeys(
  location_name = NA,
  street_address = NA,
  city = NA,
  region = NA,
  postal_code = NA,
  iso_country_code = NA,
  latitude = NA,
  longitude = NA,
  strict_address_match = FALSE,
  strict_name_match = FALSE,
  key = Sys.getenv("PLACEKEY_SECRET"),
  ...
)
}
\arguments{
\item{location_name}{The name of the place.}

\item{street_address}{The street address of the place.}

\item{city}{The city where the place is located.}

\item{region}{The second-level administrative region below nation for the place.
In the US, this is the state.}

\item{postal_code}{The postal code for the place.}

\item{iso_country_code}{The ISO 2-letter Country Code for the place.
Currently may only be US.}

\item{latitude}{Longitude in decimal degrees.}

\item{longitude}{Latitude in decimal degrees.}

\item{strict_address_match}{If set to \code{TRUE}, a Placekey is only returned if
all fields identify the place as being at the exact address specified. If set
to \code{FALSE}, the Placekey returned may ignore unit/apartment/floor information.
Optional. Default \code{FALSE}.}

\item{strict_name_match}{If set to \code{TRUE}, a Placekey is only returned if all
fields identify the POI as having the exact name specified. Optional.
Default \code{FALSE}.}

\item{key}{API key obtained from \url{https://www.placekey.io/}.
By default, loaded via environmental variable \code{PLACEKEY_SECRET}.}

\item{...}{Unused; allows \code{purrr::pmap_chr} to call \code{get_placekey}
with columns not used in the Placekey call.}
}
\value{
\code{dplyr} compatible placekeys. If address doesn't match, returns 'Invalid address'
}
\description{
Fetch placekeys using the bulk API. Automatically paginates.
}
\details{
\cr\cr Modified version #' of
\href{https://github.com/JosiahParry/placekey/blob/master/R/get_placekeys.R}{Josiah Perry's placekey package}
and bulk upload function.
}
\examples{
\dontrun{
#create a test dataframe
loc_test <- dplyr::tibble(
  location_name = rep(NA, 7),
  latitude = c(35.774323243126666, 35.7798015451112, 35.77610354072599,
               35.822304545519394, 35.81508074465238, 35.77966604109918,
               35.77385533560202),
  longitude = c(-78.63876013558352, -78.6745428292959, -78.63886492957215,
               -78.6162767007042, -78.61804304210553, -78.65351432957601,
               -78.637435700986),
  street_address = c('421 Fayetteville St', '2619 Western Blvd', '319 Fayetteville St',
                     '1205 Front Street', '2505 Atlantic Ave', '827 W Morgan St',
                     '126 E Cabarrus St'),
  city = rep('Raleigh', 7),
  region = rep('North Carolina', 7),
  postal_code = c("27601", '27606', '27601', '27609','27604', '27603', '27601'),
  iso_country_code = rep('US', 7)
) \%>\%
  dplyr::slice(rep(1:dplyr::n(), each = 1443))
#test the bulk uploader
loc_test_pk <- loc_test \%>\%
  dplyr::mutate(placekey = get_placekeys(
    location_name,
    street_address,
    city,
    region,
    postal_code,
    iso_country_code)
  )
  }
}
