% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sink_client.R
\name{init_sink_client}
\alias{init_sink_client}
\title{set up client to interact with server}
\usage{
init_sink_client(port)
}
\arguments{
\item{port}{port the server is listening on}
}
\description{
set up client to interact with server
}
\details{
the flow for use should be, in a separate process to setup
the server via run_receiving_server() and to note the
port set to pass to init_sink_client()
}

