% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kill_receiving_server.R
\name{kill_receiver_on_port}
\alias{kill_receiver_on_port}
\title{send a kill message to a listening Receiver server}
\usage{
kill_receiver_on_port(.port)
}
\arguments{
\item{.port}{port server is listening on}
}
\description{
send a kill message to a listening Receiver server
}

