% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kill_receiving_server.R
\name{kill_receiving_server}
\alias{kill_receiving_server}
\title{send a kill message to a listening receiving server}
\usage{
kill_receiving_server(.port)
}
\arguments{
\item{.port}{port server is listening on}
}
\description{
send a kill message to a listening receiving server
}

