% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the packport migration dashboard}
\usage{
launch_app(project_path = NULL, port = NULL, launch.browser = TRUE)
}
\arguments{
\item{project_path}{Optional path to pre-fill in the app}

\item{port}{Optional port to run the app on}

\item{launch.browser}{Logical. Whether to open the browser}
}
\description{
Launch the packport migration dashboard
}
