% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPM_AIC_CV_50Split.R
\name{LPM_AIC_CV_50Split}
\alias{LPM_AIC_CV_50Split}
\title{Recommended method for estimating the error variance}
\usage{
LPM_AIC_CV_50Split(XMat, yVec, numCV = 1000)
}
\arguments{
\item{XMat}{a n by p numeric matrix}

\item{yVec}{a n by 1 numeric vector}

\item{numCV}{number of cross validation splits (default is 1000)

The best way to understand how the function is to break apart the name.

\itemize{
\item LPM: Largest Plausible Model
\item AIC: we choose the largest plausible model by minimizing AIC
\item CV: we find the LPM on the training set, and get RMSE from the test set
\item 50Split: we use a 50/50 split for the training/test set.
}

This is very similar to the refitted cross validation method from \insertCite{fan2012variance}{MAIL}}
}
\description{
\code{LPM_AIC_CV_50Split} estimates the error variance with the test set RMSE across 1000 data splits
}
\references{
\insertRef{fan2012variance}{MAIL}
}
