% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Messenger.R
\name{Messenger}
\alias{Messenger}
\title{Send messages to a Reciever}
\description{
Send messages to a Reciever
}
\examples{
\dontrun{
msgr <- Messenger$new(12345) # provided from listening Reciever
msgr$send_msg("hello")
}
}

