% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate_packages.R
\name{migrate_packages}
\alias{migrate_packages}
\title{Migrate packages between R installations}
\usage{
migrate_packages(
  source_path,
  target_path,
  packages = NULL,
  dry_run = FALSE,
  upgrade = FALSE,
  ...
)
}
\arguments{
\item{source_path}{Rscript path of the source installation}

\item{target_path}{Rscript path of the target installation}

\item{packages}{Optional character vector of packages to migrate. If NULL, migrates all non-base missing/outdated packages.}

\item{dry_run}{Logical. If TRUE, return plan without installing.}

\item{upgrade}{Logical. Passed to pak}

\item{...}{Extra arguments}
}
\value{
A list with migration results
}
\description{
Migrate packages between R installations
}
