% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nested_LPM_AIC_CV_50Split.R
\name{Nested_LPM_AIC_CV_50Split}
\alias{Nested_LPM_AIC_CV_50Split}
\title{Recommended method for estimating the error variance}
\usage{
Nested_LPM_AIC_CV_50Split(XMat, yVec, numSim_Outer = 20, numCV_Inner = 30)
}
\arguments{
\item{XMat}{a n by p numeric matrix}

\item{yVec}{a n by 1 numeric vector}

\item{numSim_Outer}{number of times to select a best AIC model (default is 20)}

\item{numCV_Inner}{number of cross validation splits per selected AIC model (default is 100)

The best way to understand how the function is to break apart the name.

\itemize{
\item Nested:
\item LPM: Largest Plausible Model
\item AIC: we choose the largest plausible model by minimizing AIC
\item CV: we find the LPM on the training set, and get RMSE from the test set
\item 50Split: we use a 50/50 split for the training/test set.
}

This is very similar to the refitted cross validation method from \insertCite{fan2012variance}{MAIL}}
}
\description{
\code{Nested_LPM_AIC_CV_50Split} estimates the error variance with the test set RMSE across 1000 data splits
}
\references{
\insertRef{fan2012variance}{MAIL}
}
