% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_hub.R
\name{open_hub}
\alias{open_hub}
\title{Launch the courieR delivery hub dashboard}
\usage{
open_hub(project_path = NULL, port = NULL, launch.browser = TRUE)
}
\arguments{
\item{project_path}{Optional path to pre-fill in the app}

\item{port}{Optional port to run the app on}

\item{launch.browser}{Logical. Whether to open the browser}
}
\description{
Launch the courieR delivery hub dashboard
}
