% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packport-package.R
\docType{package}
\name{packport-package}
\alias{packport}
\alias{packport-package}
\title{packport: Cross-Platform R Package Migration Workbench}
\description{
A cross-platform R package migration workbench.
}
\author{
\strong{Maintainer}: Jane Doe \email{jane@example.com}

Authors:
\itemize{
  \item Jane Doe \email{jane@example.com}
}

}
\keyword{internal}
