% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_dispatch_log.R
\name{parse_dispatch_log}
\alias{parse_dispatch_log}
\title{Parse test log}
\usage{
parse_dispatch_log(log_path)
}
\arguments{
\item{log_path}{Path to the log file}
}
\value{
data.table
}
\description{
Parse test log
}
\examples{
tmp <- tempfile(fileext = ".log")
writeLines(c(
  "-- Failure (test-foo.R:1): addition works ----",
  "Expected 3, got 4."
), tmp)
parse_dispatch_log(tmp)
file.remove(tmp)
}
