% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_test_log.R
\name{parse_test_log}
\alias{parse_test_log}
\title{Parse test log}
\usage{
parse_test_log(log_path)
}
\arguments{
\item{log_path}{Path to the log file}
}
\value{
data.table
}
\description{
Parse test log
}
