% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_shipment.R
\name{rate_shipment}
\alias{rate_shipment}
\title{Classify shipment risk based on check and test results}
\usage{
rate_shipment(baseline_results, post_results)
}
\arguments{
\item{baseline_results}{data.table from baseline check}

\item{post_results}{data.table from post-shipment check}
}
\value{
A list
}
\description{
Classify shipment risk based on check and test results
}
\examples{
baseline <- data.table::data.table(
  severity = character(), message = character(),
  file = character(), line = character()
)
post <- data.table::data.table(
  severity = "ERROR", message = "undefined symbol",
  file = "R/foo.R", line = "10"
)
rate_shipment(baseline, post)
}
