% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sink_server.R
\name{run_receiving_server}
\alias{run_receiving_server}
\title{run a server that follows the zmq sink receiver pattern}
\usage{
run_receiving_server(cb = message, port = NULL)
}
\arguments{
\item{cb}{callback to call on message received}

\item{port}{force server to initialize on specific port}
}
\description{
run a server that follows the zmq sink receiver pattern
}
\details{
after initializing the server, will print the port it has bound to,
so that the client sending message can be set to connect
}

