% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_dependencies.R
\name{scan_dependencies}
\alias{scan_dependencies}
\title{Scan project dependencies}
\usage{
scan_dependencies(project_path)
}
\arguments{
\item{project_path}{Path to the project}
}
\value{
A data.table
}
\description{
Scan project dependencies
}
