% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sink_client.R
\name{send_message_factory}
\alias{send_message_factory}
\title{factory function to create fun to send messages to sink server}
\usage{
send_message_factory(.sink_client_list)
}
\arguments{
\item{.sink_client_list}{the list output from init_sink_client}
}
\description{
factory function to create fun to send messages to sink server
}
\examples{
# port would be set from server startup
client <- init_sink_client(5556)
send_msg <- sink_send_factory(client)
send_msg("a message")
}

