% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{wrap}
\alias{wrap}
\title{Generate a pak specification for a package}
\usage{
wrap(package, version = NULL, source_hint = NULL, github_ref = NULL)
}
\arguments{
\item{package}{Package name}

\item{version}{Optional version constraint or exact version}

\item{source_hint}{Optional hint: "CRAN", "Bioconductor", "GitHub", "local"}

\item{github_ref}{Optional GitHub ref like "owner/repo@ref"}
}
\value{
A character vector of pak specs
}
\description{
Generate a pak specification for a package
}
\examples{
wrap("dplyr")
wrap("dplyr", version = "1.1.4")
wrap("mypackage", source_hint = "Bioconductor")
wrap("r-lib/rlang", source_hint = "GitHub", github_ref = "r-lib/rlang")
}
