% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_logfile.R
\name{write_to_logfile}
\alias{write_to_logfile}
\title{write a string of text to a log file}
\usage{
write_to_logfile(x, .file)
}
\arguments{
\item{x}{string to write to file}

\item{.file}{file to write to}
}
\description{
write a string of text to a log file
}

