% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courieR-package.R
\docType{package}
\name{courieR-package}
\alias{courieR}
\alias{courieR-package}
\title{courieR: R Package Courier: Cross-Platform Delivery Workbench}
\description{
Migrates installed R packages between R installations on the same machine. Scans installed packages via manifest(), compares libraries via inventory(), and delivers packages to a new R version via ship(). Includes a Shiny dashboard (open_hub()) for point-and-click migration without requiring any packages to be installed on the origin R version.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lennon-li/courieR}
  \item Report bugs at \url{https://github.com/lennon-li/courieR/issues}
}

}
\author{
\strong{Maintainer}: Lennon Li \email{lennon.yeli@gmail.com}

Authors:
\itemize{
  \item Lennon Li \email{lennon.yeli@gmail.com}
}

}
\keyword{internal}
