% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courieR-package.R
\docType{package}
\name{courieR-package}
\alias{courieR}
\alias{courieR-package}
\title{courieR: Migrate Installed R Packages Between R Versions}
\description{
Detects all R installations on the current machine and migrates installed R packages between them. Provides find_routes() to discover R versions, manifest() to scan package libraries via subprocess, inventory() to compare two libraries, and ship() to install packages into a target R version using pak. Includes a Shiny dashboard (open_hub()) for interactive one-way and two-way migration.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lennon-li/courieR}
  \item Report bugs at \url{https://github.com/lennon-li/courieR/issues}
}

}
\author{
\strong{Maintainer}: Lennon Li \email{lennon.yeli@gmail.com}

Authors:
\itemize{
  \item Lennon Li \email{lennon.yeli@gmail.com}
}

}
\keyword{internal}
