% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courieR-package.R
\docType{package}
\name{courieR-package}
\alias{courieR}
\alias{courieR-package}
\title{courieR: R Package Courier: Cross-Platform Delivery Workbench}
\description{
A cross-platform R package courier and delivery workbench.
}
\author{
\strong{Maintainer}: Jane Doe \email{jane@example.com}

Authors:
\itemize{
  \item Jane Doe \email{jane@example.com}
}

}
\keyword{internal}
