% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_routes.R
\name{find_routes}
\alias{find_routes}
\title{Detect R installations on the system}
\usage{
find_routes(search_paths = NULL)
}
\arguments{
\item{search_paths}{Optional extra paths to search}
}
\value{
A data.frame of R installations
}
\description{
Detect R installations on the system
}
\examples{
\dontrun{
  r <- find_routes()
  print(r$version)
}
}
