% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_shipment.R
\name{inspect_shipment}
\alias{inspect_shipment}
\title{Detect project characteristics}
\usage{
inspect_shipment(project_path)
}
\arguments{
\item{project_path}{Path to the project}
}
\value{
A named list
}
\description{
Detect project characteristics
}
\examples{
\donttest{
  res <- inspect_shipment(tempdir())
  res$is_package
}
}
