% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{inventory}
\alias{inventory}
\title{Compare two package libraries}
\usage{
inventory(source_pkgs, target_pkgs)
}
\arguments{
\item{source_pkgs}{data.table from \code{manifest}}

\item{target_pkgs}{data.table from \code{manifest}}
}
\value{
A list of data.tables and a summary data.frame
}
\description{
Compare two package libraries
}
