% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{manifest}
\alias{manifest}
\title{List packages installed in a library, optionally via a different R executable}
\usage{
manifest(
  rscript_path = NULL,
  lib_path = NULL,
  format = c("data.table", "data.frame"),
  timeout_sec = 30L
)
}
\arguments{
\item{rscript_path}{Path to the Rscript executable. Defaults to current session.}

\item{lib_path}{Library path to query. Defaults to default \code{.libPaths()} of the target R.}

\item{format}{Return format}

\item{timeout_sec}{Timeout for subprocess}
}
\value{
data.table
}
\description{
List packages installed in a library, optionally via a different R executable
}
