% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_inspection_log.R
\name{parse_inspection_log}
\alias{parse_inspection_log}
\title{Parse R CMD check log}
\usage{
parse_inspection_log(log_path)
}
\arguments{
\item{log_path}{Path to the log file}
}
\value{
data.table
}
\description{
Parse R CMD check log
}
\examples{
tmp <- tempfile(fileext = ".log")
writeLines(c(
  "* checking examples ... WARNING",
  "  An example result is marked with \\\\donttest."
), tmp)
parse_inspection_log(tmp)
file.remove(tmp)
}
