% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rig.R
\name{rig_install}
\alias{rig_install}
\title{Install R via rig}
\usage{
rig_install(version, wait = TRUE)
}
\arguments{
\item{version}{R version}

\item{wait}{Logical}
}
\value{
The result of \code{\link[processx:run]{processx::run()}}.
}
\description{
Install R via rig
}
\examples{
\dontrun{
  rig_install("4.5.0")
}
}
