% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ship.R
\name{ship}
\alias{ship}
\title{Ship packages between R installations}
\usage{
ship(
  source_path,
  target_path,
  packages = NULL,
  dry_run = FALSE,
  upgrade = FALSE,
  ...
)
}
\arguments{
\item{source_path}{Rscript path of the source installation}

\item{target_path}{Rscript path of the target installation}

\item{packages}{Optional character vector of packages to ship. If NULL, ships all non-base missing/outdated packages.}

\item{dry_run}{Logical. If TRUE, return plan without installing.}

\item{upgrade}{Logical. Passed to pak}

\item{...}{Extra arguments}
}
\value{
A list with shipment results
}
\description{
Ship packages between R installations
}
