% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/take_inventory.R
\name{take_inventory}
\alias{take_inventory}
\title{Scan project dependencies}
\usage{
take_inventory(project_path)
}
\arguments{
\item{project_path}{Path to the project}
}
\value{
A data.table
}
\description{
Scan project dependencies
}
